refactor: move shared tokenserver db into tokenserver-db-common
Closes STOR-354
