+++ tokenserver-db-common/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "tokenserver-db-common"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+async-trait.workspace = true
+backtrace.workspace = true
+deadpool.workspace = true
+diesel.workspace = true
+diesel_migrations.workspace = true
+http.workspace = true
+serde.workspace = true
+thiserror.workspace = true
+
+syncserver-common = { path = "../syncserver-common" }
+syncserver-db-common = { path = "../syncserver-db-common" }
+tokenserver-common = { path = "../tokenserver-common" }
