+++ tokenserver-db-common/src/params.rs
@@ -96,30 +96,30 @@ pub struct GetServiceId {
     pub service: String,
 }
 
-#[cfg(test)]
+#[cfg(debug_assertions)]
 pub struct SetUserCreatedAt {
     pub uid: i64,
     pub created_at: i64,
 }
 
-#[cfg(test)]
+#[cfg(debug_assertions)]
 pub struct SetUserReplacedAt {
     pub uid: i64,
     pub replaced_at: i64,
 }
 
-#[cfg(test)]
+#[cfg(debug_assertions)]
 #[derive(Default)]
 pub struct GetUser {
     pub id: i64,
 }
 
-#[cfg(test)]
+#[cfg(debug_assertions)]
 pub struct UnassignNode {
     pub node_id: i64,
 }
 
-#[cfg(test)]
+#[cfg(debug_assertions)]
 pub struct RemoveNode {
     pub node_id: i64,
 }
