+++ tokenserver-db/Cargo.toml
@@ -6,14 +6,11 @@ authors.workspace = true
 edition.workspace = true
 
 [dependencies]
-backtrace.workspace = true
+async-trait.workspace = true
 futures.workspace = true
 http.workspace = true
-serde.workspace = true
 slog-scope.workspace = true
-thiserror.workspace = true
 
-async-trait = "0.1.88"
 deadpool = { workspace = true }
 diesel = { workspace = true }
 diesel-async = { workspace = true }
@@ -22,6 +19,7 @@ diesel_migrations = { workspace = true, features = ["mysql"] }
 syncserver-common = { path = "../syncserver-common" }
 syncserver-db-common = { path = "../syncserver-db-common" }
 tokenserver-common = { path = "../tokenserver-common" }
+tokenserver-db-common = { path = "../tokenserver-db-common" }
 tokenserver-settings = { path = "../tokenserver-settings" }
 tokio = { workspace = true, features = ["macros", "sync"] }
