+++ tokenserver-db/src/lib.rs
@@ -1,15 +1,12 @@
-#![allow(non_local_definitions)]
 extern crate diesel;
 extern crate diesel_migrations;
 #[macro_use]
 extern crate slog_scope;
 
-mod error;
 pub mod mock;
 mod models;
-pub mod params;
 mod pool;
-pub mod results;
 
-pub use models::{Db, TokenserverDb};
-pub use pool::{DbPool, TokenserverPool};
+pub use models::TokenserverDb;
+pub use pool::TokenserverPool;
+pub use tokenserver_db_common::{params, results, Db, DbPool};
