+++ tokenserver-db/src/mock.rs
@@ -2,12 +2,7 @@
 
 use async_trait::async_trait;
 use syncserver_db_common::{GetPoolState, PoolState};
-
-use super::error::DbError;
-use super::models::Db;
-use super::params;
-use super::pool::DbPool;
-use super::results;
+use tokenserver_db_common::{error::DbError, params, results, Db, DbPool};
 
 #[derive(Clone, Debug)]
 pub struct MockDbPool;
@@ -115,7 +110,7 @@ impl Db for MockDb {
         Ok(results::GetServiceId::default())
     }
 
-    #[cfg(test)]
+    #[cfg(debug_assertions)]
     async fn set_user_created_at(
         &mut self,
         _params: params::SetUserCreatedAt,
@@ -123,7 +118,7 @@ impl Db for MockDb {
         Ok(())
     }
 
-    #[cfg(test)]
+    #[cfg(debug_assertions)]
     async fn set_user_replaced_at(
         &mut self,
         _params: params::SetUserReplacedAt,
@@ -131,22 +126,22 @@ impl Db for MockDb {
         Ok(())
     }
 
-    #[cfg(test)]
+    #[cfg(debug_assertions)]
     async fn get_user(&mut self, _params: params::GetUser) -> Result<results::GetUser, DbError> {
         Ok(results::GetUser::default())
     }
 
-    #[cfg(test)]
+    #[cfg(debug_assertions)]
     async fn post_node(&mut self, _params: params::PostNode) -> Result<results::PostNode, DbError> {
         Ok(results::PostNode::default())
     }
 
-    #[cfg(test)]
+    #[cfg(debug_assertions)]
     async fn get_node(&mut self, _params: params::GetNode) -> Result<results::GetNode, DbError> {
         Ok(results::GetNode::default())
     }
 
-    #[cfg(test)]
+    #[cfg(debug_assertions)]
     async fn unassign_node(
         &mut self,
         _params: params::UnassignNode,
@@ -154,7 +149,7 @@ impl Db for MockDb {
         Ok(())
     }
 
-    #[cfg(test)]
+    #[cfg(debug_assertions)]
     async fn remove_node(
         &mut self,
         _params: params::RemoveNode,
@@ -162,7 +157,7 @@ impl Db for MockDb {
         Ok(())
     }
 
-    #[cfg(test)]
+    #[cfg(debug_assertions)]
     async fn post_service(
         &mut self,
         _params: params::PostService,
